Run selected set of methods and average results

Run selected set of methods and average results

Usage

deconvolute_combined(
  methyl_set,
  array = c("450k", "EPIC"),
  methods,
  scale_results = FALSE,
  ...
)

Arguments

methyl_set

A minfi MethylSet

array

type of methylation array that was used. possible options are '450k' and 'EPIC'

methods

list of methods (>1) that will be applied to the methyl set

scale_results

Whether the deconvolution results should be rescaled. Negative values will be set to 0, and the estimates will be normalized to sum to 1 per sample. Defaults to FALSE.

...

Additional parameters, passed to the algorithm used. See individual method documentations for details.

Value

dataframe with results of all selected methods as well as the combined estimates

Examples

ex_data <- minfiData::MsetEx

result <- deconvolute_combined(ex_data, methods=c('epidish','houseman'))
#> Warning: 12 NA values detected in your beta matrix. Replacing them with 0.5.
#> RPC was chosen as default for "mode"
#> blood was chosen as default for "reference"
#> Starting EpiDISH deconvolution with mode RPC ...
#> 450k was chosen as default for "array"
#> Blood was chosen as default for "compositeCellType"
#> IlluminaHumanMethylationEPIC was chosen as default for "referencePlatform"
#> IDOL was chosen as default for "probeSelect"
#> [estimateCellCounts2] The function will assume that no preprocessing has been performed. Using 'preprocessQuantile' in prenormalized data is experimental and it should only be run under the user responsibility
#> Loading required package: FlowSorted.Blood.EPIC
#> Loading required package: ExperimentHub
#> Loading required package: AnnotationHub
#> Loading required package: BiocFileCache
#> Loading required package: dbplyr
#> 
#> Attaching package: ‘AnnotationHub’
#> The following object is masked from ‘package:Biobase’:
#> 
#>     cache
#> see ?FlowSorted.Blood.EPIC and browseVignettes('FlowSorted.Blood.EPIC') for documentation
#> downloading 1 resources
#> retrieving 1 resource
#> loading from cache
#> Loading required package: IlluminaHumanMethylationEPICmanifest
#> [convertArray] Casting as IlluminaHumanMethylationEPIC
#> [estimateCellCounts2] Combining user data with reference (flow sorted) data.
#> Warning: NAs introduced by coercion
#> [estimateCellCounts2] Processing user and reference data together.
#> [preprocessQuantile] Mapping to genome.
#> Loading required package: IlluminaHumanMethylationEPICanno.ilm10b4.hg19
#> [preprocessQuantile] Fixing outliers.
#> [preprocessQuantile] Quantile normalizing.
#> [estimateCellCounts2] Using IDOL L-DMR probes for composition estimation.
#> [estimateCellCounts2] Estimating proportion composition (prop), if you provide cellcounts those will be provided as counts in the composition estimation.