Run MethylResolver

Run MethylResolver

Usage

run_methylresolver(
  methyl_set,
  doPar = F,
  numCores = 1,
  alpha = seq(0.5, 0.9, by = 0.05),
  absolute = TRUE,
  purityModel = MethylResolver::RFmodel,
  seed = 1
)

Arguments

methyl_set

A minfi MethylSet

doPar

Whether to use parallel processing to speed up the deconvolution computation if many samples are present. Default is FALSE.

numCores

Number of cores used for parallel processing to speed up the deconvolution of many samples. Requires doPar = TRUE. Default is 1. numCores = "auto" is max number of cores available minus one.

alpha

Set the alpha parameter for LTS deconvolution. This is the fraction of optimal CpGs from the signature matrix which are used for deconvolution. Must be between 0 and 1. Users can specify a vector or a single number. If a vector is specified, a grid search of the values is conducted and the alpha value that results in the lowest RMSE between the original and reconstructed mixture is selected. Default is seq(0.5,0.9,by = 0.05).

absolute

Whether to compute tumor purity and absolute cell type fractions. Default is TRUE.

purityModel

Random Forest model to predict mixture purity (unknown content) which allows the calculation of absolute cell type fractions. Required if absolute is TRUE. Default is our RF model trained on the consensus purity estimate (CPE) using TCGA data.

seed

fixed seed to account for RNG influences